fix: Delete certificate from resolution signal only if certificate is still active

+++ resolutions/signals.py
@@ -5,12 +5,13 @@ from resolutions.models import Certificate, Resolution
 
 
 @receiver(post_save, sender=Certificate)
-def certificate_delete_resolutions_images(sender, instance, **kwargs):
-    if not instance.active:
-        for res in instance.resolutions:
-            res.delete()
-        for img in instance.images:
-            img.delete()
+def certificate_delete_resolutions_images(sender, instance, created, **kwargs):
+    if not created:
+        if not instance.active:
+            for res in instance.resolutions:
+                res.delete()
+            for img in instance.images:
+                img.delete()
 
 
 @receiver(post_save, sender=Resolution)
@@ -18,5 +19,5 @@ def last_resolution_delete_certificate(sender, instance, created, *args, **kwarg
     if not created:
         if not instance.active:
             remaining_res = instance.certificate.resolutions.count()
-            if remaining_res <= 0:
+            if remaining_res <= 0 and instance.certificate.active:
                 instance.certificate.delete()
